requirements.txt
# AI Clinician with Linear Q-Learning and WDR-OPE
# Python Requirements

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Machine Learning
scikit-learn>=1.3.0
torch>=2.0.0  # For potential neural network dynamics model

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Statistics & OPE
statsmodels>=0.14.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Progress bars
tqdm>=4.65.0

# Utilities
joblib>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0

# Code quality
black>=23.0.0
flake8>=6.0.0
